pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pysisense"
version = "0.1.1"
description = "A Python SDK for interacting with Sisense API"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Sisense Ltd." }]
maintainers = [{ name = "Himanshu Negi", email = "himanshu.negi@sisense.com" }]
keywords = ["Sisense", "API", "SDK", "analytics", "BI"]
dependencies = [
  "requests>=2.32.3",
  "PyYAML>=6.0.2",
  "pandas>=2.2.3",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/sisense/pysisense"
Documentation = "https://github.com/sisense/pysisense/tree/main/docs"
Issues = "https://github.com/sisense/pysisense/issues"
Source = "https://github.com/sisense/pysisense"

[tool.setuptools]
license-files = ["LICENSE", "NOTICE"]

[tool.setuptools.packages.find]
# package code is in ./pysisense
exclude = ["tests*", "docs*", "examples*", "scripts*", "logs*", "venv-pypi*"]
